Get Marawi Barangay level data masterlist from the Open Marawi Google Drive Database

Get Marawi Barangay level data masterlist from the Open Marawi Google Drive Database

Usage

marawi_get_marawi(
  dataset = c("metadata", "demographics", "movements", "culture", "housing",
    "facilities", "establishments", "modetranspo", "conflict", "publicschools",
    "privateschools", "health")
)

marawi_get_marawi_all(tabular = TRUE)

Arguments

dataset

A string value from either "metadata", "demographics", "movements", "culture", "housing", "facilities", "establishments", "modetranspo", "conflict", "publicschools", "privateschools", and "health" specifying the masterlist data to retrieve. If not specified, will default to "metadata".

tabular

Logical. Should output be flattened into a single data.frame? Default is TRUE. If FALSE, output is a list of data.frames for each sheet in the masterlist Excel file.

Value

A tibble or a list of tibbles of the specified masterlist dataset available in the Lanao del Sur directory of the Open Marawi Database.

Author

Ernest Guevarra

Examples

## Retrieve the metadata dataset of the Lanao del Sur directory in the Open
## Marawi database
marawi_get_marawi(dataset = "metadata")
#> # A tibble: 1,201 × 9
#>    Broad Variable Categorie…¹ `Variable name` `Variable description` `Data type`
#>    <chr>                      <chr>           <chr>                  <chr>      
#>  1 Identifier                 PSGC_prov       "Philippine Standard … String     
#>  2 Identifier                 Province        "Province name"        String     
#>  3 Identifier                 PSGC_muni       "Philippine Statistic… String     
#>  4 Identifier                 Municipality    "Municipality or City… String     
#>  5 Identifier                 PSGC_bgy        "Philippine Statistic… String     
#>  6 Identifier                 Barangay        "Barangay name"        String     
#>  7 Conflict                   Grn_zero        "Barangay is one of t… dummy/bina…
#>  8 Economic                   Urban_rural     "Urban-rural classifi… dummy/bina…
#>  9 Demographics               pop_15N         "Total population per… Values     
#> 10 Demographics               hhpn_15N        "Total household popu… Values     
#> # ℹ 1,191 more rows
#> # ℹ abbreviated name: ¹​`Broad Variable Categories`
#> # ℹ 5 more variables: `Data entry` <chr>, `Sheet name` <chr>, Remarks <chr>,
#> #   `Year covered` <chr>, `Data source` <chr>
marawi_get_marawi_all()
#> # A tibble: 96 × 1,202
#>    PSGC_prov Province      PSGC_muni Municipality     PSGC_bgy Barangay Grn_zero
#>    <chr>     <chr>         <chr>     <chr>            <chr>    <chr>       <dbl>
#>  1 153600000 LANAO DEL SUR 153617000 MARAWI CITY (Ca… 1536170… Ambolong        0
#>  2 153600000 LANAO DEL SUR 153617000 MARAWI CITY (Ca… 1536170… Bacolod…        0
#>  3 153600000 LANAO DEL SUR 153617000 MARAWI CITY (Ca… 1536170… Banga           0
#>  4 153600000 LANAO DEL SUR 153617000 MARAWI CITY (Ca… 1536170… Bangco          0
#>  5 153600000 LANAO DEL SUR 153617000 MARAWI CITY (Ca… 1536170… Banggol…        1
#>  6 153600000 LANAO DEL SUR 153617000 MARAWI CITY (Ca… 1536170… Bangon          0
#>  7 153600000 LANAO DEL SUR 153617000 MARAWI CITY (Ca… 1536170… Beyaba-…        0
#>  8 153600000 LANAO DEL SUR 153617000 MARAWI CITY (Ca… 1536170… Bito Bu…        0
#>  9 153600000 LANAO DEL SUR 153617000 MARAWI CITY (Ca… 1536170… Bito Bu…        0
#> 10 153600000 LANAO DEL SUR 153617000 MARAWI CITY (Ca… 1536170… Bubonga…        0
#> # ℹ 86 more rows
#> # ℹ 1,195 more variables: Urban_rural <dbl>, pop_15N <dbl>, hhpn_15N <dbl>,
#> #   hh_15N <dbl>, hh_h_15m <dbl>, hh_h_15f <dbl>, pop_t15_0 <dbl>,
#> #   pop_t15_1 <dbl>, pop_t15_2 <dbl>, pop_t15_3 <dbl>, pop_t15_4 <dbl>,
#> #   pop_t15_5 <dbl>, pop_t15_6 <dbl>, pop_t15_7 <dbl>, pop_t15_8 <dbl>,
#> #   pop_t15_9 <dbl>, pop_t15_10 <dbl>, pop_t15_11 <dbl>, pop_t15_12 <dbl>,
#> #   pop_t15_13 <dbl>, pop_t15_14 <dbl>, pop_t15_15 <dbl>, pop_t15_16 <dbl>, …